Update patch set 2
Patch Set 2:

Here's what I got back from Luke when I asked "are you still actively maintaining selector":

"Yup and "because it just works" is correct.  (There is one outstanding bug, but it's a super corner case.)  I still use it for everything too (as do others) but have not needed to touch it in a long time.  I can do a symbolic release if needed to show that that it's still alive.  It's funny how much folks seem to value churn these days. ;-)"

I told him not to waste the symbolic release.

Patch-set: 2