Update patch set 2
Patch Set 2:

> Is there a spec on this introduction of overlapping-but-different
 > stuff in Nova?

The relevant section of the relevant spec is here:
http://specs.openstack.org/openstack/nova-specs/specs/newton/approved/generic-resource-pools.html#rest-api-impact

As the spec says, the API is going to be extracted, wholesale, to a new project so the desire is to not have dependencies on nova.wsgi & friends.

At the mitaka nova mid-cycle there was general agreement that the API should be as simple an unencumbered as possible and that we would experiment with approaches that are different from those seen thus far. The choice of selector was made (by me) because it is is a "one thing well" library, rather than a framework doing a bunch of stuff for you.

The WIP of the new API is at https://review.openstack.org/#/c/329149/ 

It's effectively plain WSGI callables, two classes with one of them using selector to dispatch to rather simple (intentionally so) module level methods. The idea is to get away from complex object hierarchies and just do the basics in a straightforward visible fashion. selector makes this easy and fairly declarative and unlike most routing libraries (including Routes) knows how to behave with appropriate HTTP.

Patch-set: 2