Update patch set 2
Patch Set 2: Code-Review-1

(1 comment)

Looking at the resolver dependancy it's only used when specifying the routes/mappings in a textfile[1] which I think would be kinda bonkers in an openstack context, so I'm kinda -0 on that.

I don't feel like I have the background on all the ins-and-outs of WSGI but it seems to be like you *can* do the new placement API with one of the already used frameworks it's just you prefer selector.  I understand not tying yourself to nova.wsgi but that doesn't mean we need another WSGI provider.  I'm -1 on adding another WSGI provider.

Also I'm -1 due to the minor issue with the upper-constraints.txt change.


[1] https://github.com/lukearno/selector/blob/master/selector.py#L195-L230

Patch-set: 2
Label: Code-Review=-1