Update patch set 2
Patch Set 2:

> I don't feel like I have the background on all the ins-and-outs of
 > WSGI but it seems to be like you *can* do the new placement API
 > with one of the already used frameworks it's just you prefer
 > selector.  I understand not tying yourself to nova.wsgi but that
 > doesn't mean we need another WSGI provider.  I'm -1 on adding
 > another WSGI provider.

Selector is not a WSGI provider, it is a routing tool. The equivalent in the existing nova stack is Routes[1] which is inscrutable and also does bad http (notably it is more than a little challenging to get it to properly produce a 405, so much so that it is not worth the effort[2]).

Nor is selector a framework. The WIP is being in a style which is sometimes called framework-less: raw WSGI applications with utilities to reduce some of the boilerplate. selector and WebOb are two utilities being used. The reasoning or goal behind framework-less is essentially transparency of HTTP and flow of control, and clear compartmentalization between resources.

Satisfying those goals is something we decided at the mid-cycle that it was worth exploring. It may be that the current strategy is no good and yes there are many alternatives (the main one in this case would be Flask because it is sort of the modern yet conservative default). That's cool, we need to have consensus on this stuff, but we need to make sure we are informed in our opinions. Maybe I should start a ML discussion on the strategy to get more eyes sooner? For the time being selector is needed for the WIPs tests to pass...

I'll fix the uc problem in a moment.

[1] https://pypi.python.org/pypi/Routes
[2] https://bugs.launchpad.net/nova/+bug/1567970

Patch-set: 2