Update patch set 3
Patch Set 3: Code-Review+1

Personally, I am in favor of the no-framework code that Chris has put together in Nova for the to-be-split-out placement API endpoint. The fact that Chris' code has no dependencies other than the (tiny) selector library is a Good Thing, IMHO, and indicates that the WSGI frameworks in Nova and other places simply aren't needed much of the time.

Best,

-jay

Patch-set: 3