Update patch set 3
Patch Set 3: Code-Review-1

OK, I spoke with Doug Hellmann on IRC a little earlier and he requested that instead of going a new direction for the placement API's WSGI code, we use one of the 3 frameworks that are in common use within OpenStack already. As good community members, we should be trying to stick with the tools in common use in the community and not deviate unless the reasons are extremely compelling, which in this case I agree with Doug, the reasons are more personal and style-related than critical to the functionality being coded in the new placement API service.

-jay

Patch-set: 3
Reviewer: Gerrit User 7 <7@4a232e18-c5a9-48ee-94c0-e04e7cca6543>
Label: Code-Review=-1