Update patch set 8
Patch Set 8: Code-Review-1

The GitHub repository appears to have Travis CI enabled for Python 3.3 and the most recent commit to the tox.ini is for Python 3.4. That said, this project seems (rather clearly) to be unmaintained. That is the sole reason for this -1.

Patch-set: 8
Reviewer: Gerrit User 12000 <12000@4a232e18-c5a9-48ee-94c0-e04e7cca6543>
Label: Code-Review=-1