Update patch set 3
Patch Set 3: Code-Review-1

Thanks for adding the explanation.

Well, no API changes and 0.7.0 is 5 years old now, so I'm fine with bumping it. But It should 1.0.1 at least (it includes bugfixes) or I'd even recommend bumping up to 1.3.0.

Why 1.3.0? upper-constraints says up to 2.0.0, but 1.3.0 is the last branch that has requirements compatible (six, pbr). 2.0.0 would require bumping also funcsigs to 1.0.0.
I'd be ok for the latter scenario, personally but 1.3.0 seems a good compromise as Debian Sid and RDO/CentOS7 have it.

If we want to be on the safe side then 1.0.1 min (which is what Debian stable ships)

Patch-set: 3