Update patch set 1
Patch Set 1:

Hi Anant, You might want to wait until https://review.openstack.org/#/c/333523/ merges.

When Keystone v3 support for sessions was merged, https://review.openstack.org/#/c/304812/,  it created a problem that we discovered a couple of days later and didn't know how to resolve. Then we ended up tagging. After we discovered the problem we reverted the change in https://review.openstack.org/#/c/331364/ and are in the process of applying a new tag, in https://review.openstack.org/#/c/333523/. 

After we get a new tag for 1.2.0, we will then get back to adding support for v3 sessions.

Patch-set: 1