Update patch set 2
Patch Set 2:

Hi Matt, 
you're right - I was applying a higher libvirt version on a stable/mitaka supported distro.

So the correct way would be to wait until the new libvirt version is officially available for this distro release and then bump the version accordingly. 

Does that make sense? If so, I'll abandon the change and keep going with private fixes for a while.

Patch-set: 2