Update patch set 2
Patch Set 2:

I'm hesitating w/ -1 as the project does not look very active to me
https://github.com/minrk/findspark

It should be simple enough to package for vendors but it does not look tested on platforms that are not OS X which could be a blocker. 
https://github.com/minrk/findspark/blob/master/findspark.py#L14
That's not very promising.

There are no tests at all ...
No indication about python3 support though, it does not look very difficult to port if required.

Patch-set: 2
Reviewer: Gerrit User 10384 <10384@4a232e18-c5a9-48ee-94c0-e04e7cca6543>
Label: Verified=0