Raise the major version of psutil
psutil is infrequently used, and mainly for agent projects like
heat-cfntools.

The reason for this change is to allow this os-refresh-config change
to call the Process.children method
Ibcbb2090aed126abec8dac49efa53ecbdb2b9b2c

The last 1.x release was 2013-11-25. 2.2.1 is available in the
following repositories making it a more appropriate version to test
against:
- EPEL 7
- Fedora 23 onwards
- Ubuntu Wily onwards

Change-Id: If31f0d5d60e8585720c4c9c95cffa202f059f6f7