Update patch set 2
Patch Set 2:

We aren't asking folks to transition from simplejson to ujson. As Monasca injests a lot of metrics in json format serialization/deserialization ends-up being one of the more important stages and contributes relatively speaking a lot to the overall cpu utilization.

We would like this library in OpenStack as a higher performance alternative to simplejson where performance is critical.

Patch-set: 2