Update patch set 2
Patch Set 2: Code-Review+2

TL;DR: +2

So currently we have json (from the std. library) and simplejson this would add a 3rd implementation ultrajson (ujson).  I spent waaaay too much time looking at the difference between ujson and simplejson, in particular how to verify that simplejson's C speedups were in effect.

The short version is I'm okay with adding this as the performance delta is a compelling reason to use ujson over simnplejson.

I'd *love* to say that we'll migrate all the OpenStack code to eitehr be on json or simplejson but I don't think that'll happen anytime soon as there are legitimate benefits to each of them.

I'd be *very* worried if we tried to add a 4th *pure* json library.

Not +W'ing this to give Matthew a chance to +/-W

Patch-set: 2
Reviewer: Gerrit User 12898 <12898@4a232e18-c5a9-48ee-94c0-e04e7cca6543>
Label: Code-Review=+2