Update patch set 1
Patch Set 1: Code-Review-2

The problem is that 2.x (and probably 3.x and 4.x) are API incompatible with 1.x.

We put the cap in place in 2014 to give people time to migrate but they haven't now we're in the situation where we can't actually do it without a shim layer somewhere (probably oslo.utils).

I'm totally in favor of raising this minimum but we can't just do it here without breaking users.

This is on the TODO list for the requirements team but it isn't high on that list :(

Patch-set: 1
Reviewer: Gerrit User 12898 <12898@4a232e18-c5a9-48ee-94c0-e04e7cca6543>
Label: Code-Review=-2