Update patch set 2
Patch Set 2:

What's the compatibility with Django versions?

One of the difficulty that distributions have is that BaseOS ships a different version that the ones Horizon is supposed to be compatible with. We've worked with Horizon developers to have a "stable" django version for Horizon to lessen the burden (currently, it's 1.8.x)

I'd appreciate if you can answer that.

Patch-set: 2