Bump 'global-requirements' for 'oslo.concurrency' to 2.6.1
In this[1] change, when it was initially proposed to update
'global-requirements', but we ended up only updating the
'upper-constraints.txt', because of concerns that bumping
'oslo.concurrency' also bumped related components: 'pbr', 'oslo.config'
and 'oslo.utils'.

Another argument went:

    "We do not generally raise the minimum version of dependencies to
    fix bugs, and we especially don't want to do that in stable
    branches. Please update this to only changes the
    upper-constraints.txt file."

However, these above arguments are questioned in this Nova backport[2],
where we try to handle the case of deployments not having new
enough oslo.concurrency.

All of this to get a fix for CVE-2015-5162 -- "qemu-img calls need to be
restricted by ulimit"

[1] [stable/liberty requiements]
    https://review.openstack.org/#/c/333403/ -- Bump 'upper-constraints'
    for 'oslo.concurrency' to 2.6.1

[2] [stable/liberty openstack-nova]
    https://review.openstack.org/#/c/327624/ -- virt: set address space
    & CPU time limits when running qemu-img

Depends-On: I2bc4190ab6996b22e56095b1a653ebe4cb3b6513

Change-Id: Ic9f75f2147c53da72e3aa0b2369c5f8de4d8d435