Update patch set 1
Patch Set 1: Code-Review-2

This is okay from a testing POV and I agree that in themselves each requirement has been approved from a stable backport POV.  We've run 2.6.1 lightly in our gate and it works. huzzah!

What this does is raises the minimum requirement on a number of packages which in turn raises the minimum *and* adds new packages.  This is an unfair burden to place on our distribution partners.  If they choose to stay on $older packages that are within the range supported by global-requirements that's fine we accept that.  They're also choosing to take the risk on the associated CVE.

Yes we *did* raise the minimum on mitaka but the impacts were much smaller.

Patch-set: 1
Reviewer: Gerrit User 12898 <12898@4a232e18-c5a9-48ee-94c0-e04e7cca6543>
Label: Code-Review=-2