Update patch set 1
Patch Set 1:

I don't think that's a real burden on distributors. Dealing with bumped deps and occasional new packages is a standard part of their job which they are all well practised at. The alternative approach is that Nova checks the version at startup and aborts nova-compute if the new enough oslo.concurrency isn't available, which is worse because its the same result in a manner which distributors don't see upfront.

Patch-set: 1