Add hiredis-py to requirements
Is the library actively maintained?
Yes. The lastest commit is about a week ago and the latest release is
about a year ago. Pls refer to https://github.com/redis/hiredis-py

Is the library good code?
Yes, it has a good code style and UT.

Is the library python 3 compatible?
Yes.

Is the library license compatible?
Yes, BSD license.

Is the library already packaged in the distros we target (Ubuntu latest
/ Fedora latest)?
Yes for Ubuntu, No for Fedora

Is the function of this library already covered by other libraries in
global-requirements.txt?
No

Is the library required for OpenStack project or related dev or
infrastructure setup? (Answer to this should be Yes, of course) Which?
Yes, Dragonflow, pls refer to https://review.openstack.org/#/c/336351/

Change-Id: Iae90fb647deed3000c8d2a4a3cbabc3f18fec790