Update patch set 1
Patch Set 1:

Granted we are at 2.6.0 in ubuntu so we could easily get to 2.6.1. However I'm not sure what the case is for other distros.  I know that if we were at 2.3.0 then this change would force us to backport the 2.6.1 changes to 2.3.0 ourselves because we can only bump to stable point releases after final release of ubuntu/openstack.

Patch-set: 1