Reset oslo.utils to pre-GA liberty version constraint
At liberty GA 1cc88e0d007acb49d4fa2f7633ed1302b6d2bd4a the
oslo.utils upper-constraint was 2.5.0. We've rolled back some
automatic updates to u-c in liberty which were a mistake and
stopped in I97f0cf78bf2c980628abe0f18aa9f5e96af70ae7.

Change Ic9f75f2147c53da72e3aa0b2369c5f8de4d8d435 proposes
raising the minimum required oslo.concurrency>=2.6.1 as a
dependency for a nova CVE bug in stable/liberty, but doing so
would pull in some updated versions of other packages, one of
which is oslo.utils. So to avoid pulling in things from oslo.utils
that were meant for mitaka, fix the cap for oslo.utils in liberty.

This also fixes the cap for debtcollector and stevedore which
are dependencies of oslo.utils, so if we're going to change the cap
for oslo.utils we should do the same for it's dependencies.

Change-Id: I2bc4190ab6996b22e56095b1a653ebe4cb3b6513
Related-Bug: #1449062