Update patch set 5
Patch Set 5:

I'm leaving my -2 where it is for now.

I feel like the impact on *other* services i(all consumers of oslo,concurrency) is evidence that we're taking a larger risk here than I think is being credited.

I *will* remove my -2 once I'm confident we've looked at the other options.

The fact is if we land this we're forcing a minor version bump on *every* consumer of oslo.concurrency on the liberty branch.

Patch-set: 5