Update patch set 5
Patch Set 5: -Code-Review

I'd still be fine with updating the g-r as a distro packager (as we target UC).  Though from a stable-maint perspective this is a larger change with knock-on effects we should avoid.  From a stable-maint perspective we should try/push hard on getting the fix backported to 2.3.x.  I realize this could be hard, but it is still technically possible and worth exploring.

Patch-set: 5