Update patch set 5
Patch Set 5:

I'm still -1 until this is the last option available.  It seems like backporting to 2.3.1 should be a possibility.  2.3.0 has been a supported version for Liberty since it released.  It would be much less intrusive to bump the min in g-r to 2.3.1.

With that said, from a distro pov we're bumping to 2.6.1 in ubuntu since we're already at 2.6.0.

Patch-set: 5