Update patch set 1
Patch Set 1: Code-Review-1

the global-requirements sync needs to be configured for that project first as well as merging the missing requirements. 

I overall welcome the addition though, the -1 is only because of the missing depends-on: for the corresponding infra change.

Patch-set: 1
Reviewer: Gerrit User 6593 <6593@4a232e18-c5a9-48ee-94c0-e04e7cca6543>
Label: Code-Review=-1