Update patch set 4
Patch Set 4:

Just an update, the requirements team is looking into a 'optional-requirements.txt' file for stuff like this.  Things that are not used directly but could be useful for various reasons (performance) and that are optional.

Having just said that, this sounds like automagic dependencies...  Here's a link https://wiki.gentoo.org/wiki/Project:Quality_Assurance/Automagic_dependencies

The basics are that changing software behavior is bad and hard to test.  There's also the issue that if one project encourages an optional dep it could cause issues in other projects (perhaps the plugin has a flaw).

Patch-set: 4