Update patch set 1
Patch Set 1: Code-Review-1

This package seems to embed its own list of ROOT CA. This is a no-go. The list of certificate should be the ones used system-wide. In the case of Debian / Ubuntu, that would be the ca-certificates package, which install stuff in /usr/share/ca-certificates/mozilla.

How can we even be sure that the list of ROOT CA in this package is accurate, trustable and up-to-date? I wouldn't trust it, no mater what.

Patch-set: 1