Update patch set 5
Patch Set 5:

> We've reached an impasse.  What is the status here?

Given it's my -2 that's causing the impasse I guess the status is I need to re-evaluate exactly what this change does.

While the U-C changes are correct IMO they don't change the bottom line that this change is *forcing* distributions to take the hit on several major version increases in oslo packages (mostly due to the removal of the olso. namespace).  Our testing shows that that isn't a problem but it really only tests defcore projects.  It also adds at least one new package.

We did choose to bump the minimum on mitaka but the delta was smaller and more contained.  From a release POV it has impacts on the release management teams and *every* project team that require oslo.concurrency that's 30 for liberty and 35 for mitaka

We've had a long standing line commitment that we won't raise these minimums without good reason, while a CVE is certainly compelling I'm not (yet) convinced it is adequate motivation.  

The nova change is passing CI with (untested) fallback code

Frankly I'm very conflicted as to which of the options is the least bad.

Patch-set: 5