Update patch set 2
Patch Set 2:

To be more clear, it sounds like the commit message is saying that once that is merged we can remove the block.  I don't think that is the case, but the commit message could be more clear.  Just need to remove the 'until' and say that we need to block the broken release.

Patch-set: 2
Label: Verified=0