Prevent setuptools from upgrading to 25.0.0
In the 25.0.0 release [1] of setuptools during any install
operation the package in not overwritten. This is a temporary
solution to unblock gate jobs. If a package is installed from
another requirement via pip and then is installed again
from git, it is not updated causing check_libs_from_git in
devstack to fail.

We can either:
* set SETUPTOOLS_SYS_PATH_TECHNIQUE=rewrite
* uninstall packages before reinstalling them, to ensure that
  they're install from git

[1] https://github.com/pypa/setuptools/commit/e211f4584ac59f404a3e1a17c6b54dff18468cf6

Change-Id: Ie8f7fe6f0278821e82e39afea8a3f84f66cde4ca
Related-Bug: #1605998