Update patch set 1
Patch Set 1: Code-Review-1

This will result in a bunch of comment only changes in *all* managed projects that use an affected package.

I'm all for consistency but we need to time this to reduce the pain on consumers.

I'd be find with updating the comment the *next* time we change the requirements spec

Patch-set: 1
Reviewer: Gerrit User 12898 <12898@4a232e18-c5a9-48ee-94c0-e04e7cca6543>
Label: Code-Review=-1