Update patch set 1
Patch Set 1: Code-Review-1

Looks at project config and devstack[1] The 2 server versions we install are 1.7.5 and 2.3.2[2].  I note that you tested 1.7.3 and and 2.3.3.  Can you verify that 1.7.5 and 2.3.2 work with client versions 1.9.0 and 2.3.0

I'm -1 without that confirmation.  I'll add the PTLs for Monasca, Freezer, Fuel, and Ceilometer to this review for awareness as we don't want to "blindside" them.

[1] I know there is some contention there ... I'm ignoring that g=for now.
[2]

 jenkins/jobs/macros.yaml:856:          wget https://download.elastic.co/elasticsearch/elasticsearch/elasticsearch-1.7.5.deb
 jenkins/jobs/macros.yaml:857:          sudo dpkg -i elasticsearch-1.7.5.deb
 jenkins/jobs/macros.yaml:869:          wget https://download.elastic.co/elasticsearch/release/org/elasticsearch/distribution/deb/elasticsearch/2.3.2/elasticsearch-2.3.2.deb
 jenkins/jobs/macros.yaml:870:          sudo dpkg -i elasticsearch-2.3.2.deb

Patch-set: 1