Update patch set 4
Patch Set 4: Code-Review+2

It seems reasonable to me to add this.  I'm not a big fan of vendoring but IMO that isn't a reason to down vote this.  The setup.py is strange but ....

Not +W to give Matthew one last chance to say "no".

Patch-set: 4
Reviewer: Gerrit User 12898 <12898@4a232e18-c5a9-48ee-94c0-e04e7cca6543>
Label: Code-Review=+2