Require oslo.rootwrap>=5.0.0
rootwrap 5.0.0 provides the ability to run privsep without additional
config, which is needed to seamlessly rolling forward on privsep
enabled os-brick, os-vif, nova, neutron, cinder.

We bump upper-constraints and raise the minimum required version in
g-r to make grenade work. Without bumping the minimumm, the version of
rootwrap from mitaka satisfies master and we don't get the fix on
upgrade (to use privsep-helper).

It also unpins os-brick 1.4.0, which was blocked only because of the
use of privsep and no upgrade path.

Depends-On: I3a52f762deb176fe9201b2a0f0da363057f8aaec

Change-Id: Ic2fc55fd33e35f26a06f8655224af12b705dca4e