Update patch set 1
Patch Set 1:

> Should probably rebase on https://review.openstack.org/#/c/350192/

Thanks for the review, Eric. The rebase is pretty ugly as it pulls in dozens of patches to be rebased on. I dunno if we have to wait or otherwise. Hopefully, this patch works as is and that I can leave upto the reviewers to merge rightfully.

Patch-set: 1