Add ndg-httpsclient to global-requirements
Ansible makes use of httplib/urllib for the get_url task. The package
ndg-httpsclient provides the appropriate improvements to support SNI
extensions for SSL certificates which are becoming more commonplace,
causing failures for this task.

Having this package in place along with pyopenssl and pyasn1 will
also get rid of the SNIMissingWarning issued by requests' vendored
urllib3.

The commit history shows development activity in 2016:
https://github.com/cedadev/ndg_httpsclient

I can't speak to whether it's good code.  That is outside my skill
set.

Based on the distribution packages available for it, it does appear
to be Python3 compatible.

The license is BSD, which is compatible.

The package is included in the following distributions:
 - Ubuntu [1] (Xenial only)
 - RHEL/CentOS [2]
 - Fedora [3]
 - Gentoo [4]
 - SuSE [5]

[1] http://packages.ubuntu.com/search?suite=all&searchon=names&keywords=ndg-httpsclient
[2] https://www.rpmfind.net/linux/rpm2html/search.php?query=python-ndg_httpsclient
[3] https://apps.fedoraproject.org/packages/python-ndg_httpsclient
[4] https://packages.gentoo.org/packages/dev-python/ndg-httpsclient
[5] https://build.opensuse.org/package/show/devel:languages:python/python-ndg-httpsclient

Change-Id: I7f90ab6779944bb7e3c9ba4b62fe33058cec6af4