Update patch set 4
Patch Set 4:

While I don't like that we'd need to add yet another client I can see the need of it.  Not having a sni compatible client is a pain otherwise.  I'd like other cores to talk about it before merging.

on another note, this would be a good canidate for optional-requirements

Patch-set: 4
Label: Verified=0