Add dogtag-pki used by Barbican
- Is the library actively maintained?

Latest version is 10.3.5.0.1 released on 2016-08-08. Maintainers
are responsive in #dogtag-pki on freenode.
Also RedHat is using it for the "Red Hat Certificate System" product.

- Is the library good code?

I haven't looked into the code in detail but is has a testsuite (using tox
and pytest). In general, the code looks good to me.

- Is the library python 3 compatible?

Yes

- Is the library license compatible?

License is LGPLv3+ which is compatile.
Note: The library was relicensed *because* of the OpenStack license
requirements.

- Is the library already packaged in the distros we target (Ubuntu
latest / Fedora latest)?

Ubuntu:
Not yet. But given that it's a common python package it shouldn't be hard to
package it for Ubuntu.

Fedora:
Not yet. But given that it's a common python package it shouldn't be hard to
package it for Fedora.

Note: Both (Ubuntu and Fedora) have the server part of dogtag already packaged.

- Is the function of this library already covered by other libraries
in ``global-requirements.txt``?

No.

- Is the library required for OpenStack project or related dev or
infrastructure setup? (Answer to this should be Yes, of course)
Which?

Yes. It's used for the dogtag Barbican plugin.
See http://git.openstack.org/cgit/openstack/barbican/tree/barbican/plugin/dogtag.py

Change-Id: Ibedae54e631e9c3d3726453adcd204ce96b19d77