change django_openstack_auth to not use canonical name
The release job that tried to update this constraint failed because the
name in the setup.cfg package metadata did not match the name here in
the constraints list. We could change the name in the metadata, but that
would break a bunch of links to existing build artifacts. Since a lot of
other packages in this file do not use the canonical name, let's change
this entry to use the name from setup.cfg instead of the canonical form.

Change-Id: I2a584195063fed11b4ee51b783045331815a8576
Signed-off-by: Doug Hellmann <doug@doughellmann.com>