Bump keystoneclient to 2.0.0
Latest version is 3.4.0

keystoneclient < 2.0.0 has a bug related to discovery keystone version
while using session object. It was fixed in 2.0.0 with
  https://github.com/openstack/python-keystoneclient/commit/d9031c252848d89270a543b67109a46f9c505c86

You can name it as a feature for keystoneclient >= 2.0.0 instead of a bug for
keystoneclient <2.0.0 , but it will not change nothing: interface of old
keystoneclient is different from new one.

Change-Id: I6cb1a4646a03715fedaa5075c9c3aedcf2d7184f