Remove vitrage from projects
Vitrage is using a library not tracked in GR or UC (exrex).  This
library is licenced AGPLv3, but used only for testing, so it may
end up being allowed to be added.  However, upstream vitrage has
not responded to our bug requesting action on this so I move to
remove them from projects.txt

Bug: 1612352
Change-Id: I981599cea4d9269fb1831c96a4433dcc5e683a57