Update patch set 1
Patch Set 1: Code-Review+1

> Looks good to me .. but
 > @Nikhil: Requirement: python-glanceclient!=2.4.0,>=2.0.0 (used by
 > 80 projects)
 > 
 > You can see that cinder, glance, horizon, keystone, neutron, nova
 > and swift are okay with this
 > 
 > Do you need extra testing before we merge this?

Thanks for your review Tony. I wanted to push for this as 2.5.0 has a revert for a regressing 2.4.0 introduced. However, there's a commit in 2.4.0 about warlock model creation that helps a subset of the community.

I am positive about this release to work as the regression in 2.4.0 has been tested and reverted herein. In the worst case, we will have to revert to the old constraint of 2.3.0 but less likely.


https://review.openstack.org/#/c/358391/1

Patch-set: 1
Reviewer: Gerrit User 2537 <2537@4a232e18-c5a9-48ee-94c0-e04e7cca6543>
Label: Code-Review=+1