Bump g-r python-glanceclient to 2.5.0
As per the discussion in the Glance meeting today [1], it was brought up
that due to requirements sync in python-glanceclient since 2.3.0 we
should keep the minimum requirement to 2.5.0. This keeps a clean
perspective on the release's dependency requirements and avoids having
to keep a negative flag (i.e. !2.4.0) in the file.

Also, release 2.5.0 has been in the gate for couple of days and can be
assumed to be safe for this bump.

[1] http://eavesdrop.openstack.org/meetings/glance/2016/glance.2016-08-25-14.00.log.html#l-113

Change-Id: Ibb5e39369a34ce1d34df34472034d9f6c72fff11