Update patch set 1
Patch Set 1: Code-Review-2

The cap on sphinx was added in I63c6159eb23430a181eb20016b35366c2fbbf9f8 because it 'breaks things'.

Sadly there isn't enough evidence to that the review pointed at in the commit message is an exhaustive list of 'things'

Removing this cap will allow a range of upper-constraints changes that are difficult to verify in the gate.  I'd be much more comfortable taking this patch after we branch stable/newton.

Patch-set: 1
Label: Code-Review=-2