Update patch set 2
Patch Set 2:

Starting with os-brick 1.5.0, command execution was switched over from using rootwrap to using privsep.

Privsep requires some additional initialization that was not present in mitaka. I don't think we want to backport everything to mitaka to enable that either.

Is there something in os-brick 1.5 or 1.6 that is needed in Mitaka?

Patch-set: 2