Interim update to requirements for Newton
This is an interim update to requirements for newton.

Multiple changes in Newton require a new client and these include
upgrades, force deleting instances, support for flavor id's that start
with a leading zero, and additional information in flavor command. The
ability to schedule a backup requires the new client, and the new
client exposes a quota update API as well. backup-create has a new
--incremental flag.

For these reasons, I'm increasing both minimum and u-c.

Change-Id: I3e469b0358d7371cf69c2be9015ee79e36c8c0f8
Depends-On: I12cf041cc877fe00fbdc998bfcd4bb16c089d2b3