Bump pyghmi version to 1.0.9
The pyghmi version 1.0.9 fixes an annoying problem in the IpmiServer
implementation where different ipmi clients could try to reuse a UDP
port which was already being used, causing some requests to error out.

Projects like virtualbmc and openstack-virtual-baremetal does rely on
the IpmiServer implementation and are used to in the OpenStack gate for
Ironic and Tripleo respectively. Bumping pyghmi to a newer version where
this problem has been fixed does help on having a more stable gate.

Change-Id: I2840e785ef1e62be267402b276607ee06e077bef