Bump cryptography to work with patched openssl
cryptography embeds some data structures and function signatures from
openssl in its source code. One of these changed in a recent openssl
security release. The result of this is that it is not possible to
build python-cryptography<1.2.3 against the latest security patch of
openssl. We're pinning it to 1.1.2 for liberty, which is affected.

The issue is described here:
https://github.com/pyca/cryptography/issues/2750

There are no 1.1.X releases after 1.1.2. I have confirmed that 1.2.3
is buildable, but 1.2.2 is not, so 1.2.3 is the minimum supportable
version.

The cffi bump to 1.6.0 is a dependency, as cryptography 1.2.3 requires
def_extern from cffi, which was not available in 1.3.1.

Change-Id: Iec0cf800f612213089e4997780473c5b6f9ae1d7