Block osc 3.2.0
osc 3.2.0 goes into a blocking auth prompt in our multinode
environment (sometimes) which is blocking a bunch of jobs. It's
extremely unclear why that's the case. There is no timeout for this
prompt, so it holds until the system times out at 3 hours.

We should block this, and revisit in the next release what is going on
here.

Change-Id: I8198d36edd7a3c226be1b0a696c0e9eeb4a2b086