Update patch set 1
Patch Set 1: Code-Review+1

> I think Neutron wanted this for Newton to be in upper-constraints,
 > but I'll leave it to them to post the requirements freeze exception
 > request in the ML.

The bug lies with os-vif and Nova, that uses the library. It'd be nice to have it to address gate failures, but strictly speaking we're good on our side because we addressed our side of the race already.

Patch-set: 1
Reviewer: Gerrit User 748 <748@4a232e18-c5a9-48ee-94c0-e04e7cca6543>
Label: Code-Review=+1