Update patch set 2
Patch Set 2: Code-Review-1

OK, *this* patch would need to bump oslo.db to >13.3. Though I wonder what's the real need to push for such a major bump right now when we are in reqs freeze mode. I believe that a mere block of pymysql 0.7.7 and oslo.db .13.[23] is a much better isolated fix for Newton.

Patch-set: 2