Update patch set 2
Patch Set 2:

I suppose I formatted the patches in a diferent manner, each patch is targeted to a dependancy, not an issue.  At this point I have to wait on 4.13.3 anyway so when that comes out I can change the patch ordering.

I do not think it's a good idea to submit a series of patches that 'patch a patch' though, which I was trying to avoid.  But here it's likely valid.

Patch-set: 2