[DNM] Sample lower-constraints.txt
This ia actually a lower-constraints.txt which is pretty simple to generate.

There are a few lower bounds that don't even exist on pypi anymore :(

 - PuLP: 1.0.4 not available changing to 1.4.1
 - ldap3: 0.9.8.2 not available changing to 0.9.8.7.post1
 - PrettyTable: 0.7 not available changing to 0.7.1
 - pyldap: 2.4 not available changing to 2.4.20
 - pyngus: 2.0.0 not available changing to 2.0.2
 - PyYAML: 3.1.0 not available changing to 3.10
 - rtslib-fb: 2.1.41 not available changing to 2.1.43
 - XStatic-Bootstrap-SCSS: 3 not available changing to 3.1.1.1
 - XStatic-jQuery: 1.7.2 not available changing to 1.8.2.1
 - XStatic-jquery-ui: 1.10.1 not available changing to 1.10.4.1
 - XStatic-term.js: 0.0.4 not available changing to 0.0.4.1

Also the minimum versions for:
 - simplejson
 - xattr
 - cliff-tablib
 - testresources
 - couchdb

Don't build under py3.  Rather than split the minimum based py pythn_version
it's better to just pick a version for each that works in both environments.

Change-Id: I6732d733df9f72b4414cce4244975394aae33a53