Add murano-pkg-check 0.1.0
Murano project has a blueprint about creating tool for validation
of application packages [1]. It was implemented as a separate library,
because it can be also used separately from Murano code base.
In other hand, it is necessary to have this library in the global
requirements to allow Murano project use it

* Is the library actively maintained?
Yes; this library is mantained by Murano team

* Is the library good code?
Yes; it is good as Murano project code

* Is the library python 3 compatible?
Yes, it has py34 gate [2].

* Is the library license compatible?
Yes Apache 2.0 lic.

* Is the library already packaged in the distros we target
(Ubuntu latest / Fedora latest)?
No.

* Is the function of this library already covered by other
libraries in global-requirements.txt?
No

* Is the library required for OpenStack project or related dev
or infrastructure setup? (Answer to this should be Yes, of course)
Which?
Yes, murano-pkg-check is required for Murano project. See this
patchset [3]

1. https://blueprints.launchpad.net/murano/+spec/package-validation
2. https://github.com/openstack-infra/project-config/blob/master/zuul/layout.yaml#L10988
3. https://review.openstack.org/#/c/357723/

Change-Id: I992d4d550ca81f77feea87a34c1352e6f881baf7