Update patch set 1
Patch Set 1:

> If the procedure is to keep things broken, then the procedure is
 > wrong. I raised that concern many times. This is supposed to be for
 > downstream distros, and downstream distros (ie: Debian that I
 > represent, Ubuntu, and RedHat) agree with me.

It's a question of timing.  Our release timeline is well documented.  This change could have happened at pretty much anytime in the last 6months it just arrived too late.

Yes distros are part of the rational but only part.  I'd be happy to talk to you about this in Barcelona

Patch-set: 1