Update patch set 1
Patch Set 1:

As this is the first adding of the package can you please answer the questions listed here?  https://github.com/openstack/requirements/#for-new-requirements

You are also missing the addition of it to upper-constraints.txt

Also, kuryr-libnetwork depends on this (along with kuryr-kubernetes I think from the email).  Do other projects depend on those (are there possible knock on changes)?

Patch-set: 1
Label: Workflow=0