Update patch set 5
Patch Set 5:

@tony, et al: While I understand you are still feeling conflicted, This has been stuck in limbo for ages, and will only just rot away, until somebody comes to abandon this.

We've had enough time (2-ish months) to mull over this, now we ought to make progress and bring this to a logical conclusion.


So, we currently have two options

(1) Bump global-requirements (this patch)

(2) Or merge this backport -- https://review.openstack.org/#/c/327624/ -- "virt: set address space & CPU time limits when running qemu-img"

That also adds a check for the presence of 'ProcessLimits' attribute (which is only present in oslo.concurrency>=2.6.1; and a conditional check for 'prlimit' parameter in qemu_img_info() method.

Patch-set: 5