Add sphinxmark for doc publishing
The sphinxmark package is already used by openstack-manuals and
we would like to also use it in OpenStack-Ansible in order to
clearly mark the release name in published documentation. This
reduces confusion about which release the documentation applies
to.

The package is only used in test-requirements.txt for the purpose
of documentation publishing.

Change-Id: Iefab37db255da0b1aadef64a55b0a519c489908c