Update patch set 1
Patch Set 1: Code-Review-1

(2 comments)

I'd consider taking another look at the effects of normalizing the global-requiremnets file.

Last time we saw a large number of whitespace only changes which we want to avoid.

We should also decide which the canonical format is and use that in both places to make the this scenario go away.

Patch-set: 1
Label: Code-Review=-1