Update patch set 1
Patch Set 1:

We are in the process of merging code from monasca-statsd master branch to newton. The tags used to create the newton branch were a little old. If we get the newton branch updated with what is currently on master branch, and that fixes the problem, would that be acceptable?

Patch-set: 1