Update patch set 1
Patch Set 1:

Hi Anita, I've cherry-picked a couple of commits from the master monasca-statsd branch to the newton branch that should resolve the issues in requirements.txt. Please see the review at, https://review.openstack.org/#/c/374361/, which updates the commit ID for the newton release of monasca-statsd. If that review is merged, then I don't believe this review should be merged, and should be abandoned if I understand correctly. Please let me know if there is anything else that I'm missing or needs to be resolved. Thanks for all your helps so far. --Roland

Patch-set: 1