Update patch set 2
Patch Set 2:

> I do see at least one remaining problem.  We need support for
 > python 3.4 AND 3.5 as both Ubuntu Trusty (3.4) and Xenial (3.5) are
 > supported.

Hi Matt, we have resolved this problem in the new patch. We added python 3.5 support in the new version of abclient on PYPI, and also have tested it on both Ubuntu Trusty (3.4) and Xenial (3.5). We also tested it with tox supportting python 3.4 and 3.5.

Hope you could help us to check it again. Thanks.

Patch-set: 2