Update patch set 2
Patch Set 2: Code-Review-1

The setup.py doesn't mention the licence or supported python versions.  I'm not sure if it actually supports python3 based on the use of naked print statements in cli.py (line 56 and 69).  The package also does not define dependencies (test doesn't need any as it's all built in it looks like, same with ofpstr itself), not sure if it's best to define empty sets here or not.  I didn't do a deep dive on the code quality, but it looks to be a small library/client...

Patch-set: 2
Reviewer: Gerrit User 14288 <14288@4a232e18-c5a9-48ee-94c0-e04e7cca6543>
Label: Code-Review=-1