Update wsgi-intercept to 1.4.1
1.4.1 adds (optional) support for checking that response headers
sent by the intercepted application are in the proper form (`str`
objects in python 2 and python 3, thus bytes in one and not the
other). This is hard to test in the code that uses wsgi-intercept
because most HTTP libraries will normalize the response headers
to a convenient type.

This functionality will be used in in the tests of the placement
API in nova.

This patch also changes the wsgi-intercept entry in global-requirements
to use the version of the name with a '-' not an '_' so that there is
consistency between the global-requirements and upper-constraints files.

Change-Id: I3b8aabda929fe39b60e645abb6fabb9769554829