Update patch set 3
Patch Set 3:

The delta from 1.1.2 to 1.2.3 is a little scary for a liberty release.

Looking at distros that liberty was targeted at 1.0.2g isn't packaged so either the security fix has been silently back ported or wasn't deemed severe enough for backporting.

I've requested a 1.1.3 release with *just* the fix(es) needed to build with 1.0.2.

Patch-set: 3