Update patch set 2
Patch Set 2:

There is not a bug for nova-lxd, no. We discovered the issue while trying to get a clean tempest test run.

pylxd is only used by nova-lxd, afaik, and I'm also the upstream for pylxd. Someone else added it to requirements, so I'm trying to make sure that stays up to date.

Patch-set: 2