Update patch set 2
Patch Set 2: Code-Review-1

Matthew,

I thought so based upon irc chatter, but I misunderstood the communication.  I am changing my vote back to -1.  I will personally investigate this tomorrow morning and provide a report back by 10am pst.

We had an upstream bug that gave us a false positive at exactly the same time as swapnil tested this change, so we want to triplecheck it.  Worse comes to worse, we can release with this change and sort it out in kolla directly since we have our own docker interface module (between ansible and docker-py).  I guess there really is no need to block on Kolla's -1 - its more a way for me to keep track of this issue (using gerrit.  This probably isn't appropriate use of gerrit or the requirements team's attention.  Apologies for that.  We are really struggling to keep up with our workload, especially with a ptl transition happening at the same time.  Let me know how you wish to proceed.  I hit the wall a couple hours ago so i'm out of steam for tonight.  Swapnil has +2ed the change,but I think he is waiting on validation himself (the triplecheck thing I mentioned).

Swapnil, if you have time and since your in a different timezone, an easy test would be to merge your patch and my rabbitmq hack - if that passes the gate, we should be good to go as our gate tests upgrade, deploy, and reconfigure which provide good coverage for all of the various kolla docker interfaces.

Hitting the rack :)  night all.

Patch-set: 2